refactor AddCommand and AddCommandParser into a separate package for property
