+++ src/main/java/seedu/condonery/logic/parser/CondoneryParser.java
@@ -6,7 +6,6 @@ import static seedu.condonery.commons.core.Messages.MESSAGE_UNKNOWN_COMMAND;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
-import seedu.condonery.logic.commands.AddPropertyCommand;
 import seedu.condonery.logic.commands.ClearCommand;
 import seedu.condonery.logic.commands.Command;
 import seedu.condonery.logic.commands.DeleteCommand;
@@ -15,7 +14,9 @@ import seedu.condonery.logic.commands.ExitCommand;
 import seedu.condonery.logic.commands.FindCommand;
 import seedu.condonery.logic.commands.HelpCommand;
 import seedu.condonery.logic.commands.ListCommand;
+import seedu.condonery.logic.commands.property.AddPropertyCommand;
 import seedu.condonery.logic.parser.exceptions.ParseException;
+import seedu.condonery.logic.parser.property.AddPropertyCommandParser;
 
 /**
  * Parses user input.
@@ -26,7 +27,7 @@ public class CondoneryParser {
      * Used for initial separation of command word and args.
      */
     private static final Pattern BASIC_COMMAND_FORMAT = Pattern
-        .compile("(?<commandWord>\\S+)\\s(?<directoryArgument>-[pc]{1})(?<arguments>.*)");
+            .compile("(?<commandWord>\\S+\\s-[pc]{1})(?<arguments>.*)");
 
     /**
      * Parses user input into command for execution.
@@ -42,13 +43,12 @@ public class CondoneryParser {
         }
 
         final String commandWord = matcher.group("commandWord");
-        final String directoryArgument = matcher.group("directoryArgument");
         final String arguments = matcher.group("arguments");
-
+        
         switch (commandWord) {
 
         case AddPropertyCommand.COMMAND_WORD:
-            return new AddCommandParser(directoryArgument).parse(arguments);
+            return new AddPropertyCommandParser().parse(arguments);
 
         case EditCommand.COMMAND_WORD:
             return new EditCommandParser().parse(arguments);
