+++ src/main/java/seedu/condonery/logic/parser/property/AddPropertyCommandParser.java
@@ -1,4 +1,4 @@
-package seedu.condonery.logic.parser;
+package seedu.condonery.logic.parser.property;
 
 import static seedu.condonery.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
 import static seedu.condonery.logic.parser.CliSyntax.PREFIX_ADDRESS;
@@ -9,7 +9,12 @@ import java.util.Set;
 import java.util.stream.Stream;
 
 import seedu.condonery.logic.commands.Command;
-import seedu.condonery.logic.commands.AddPropertyCommand;
+import seedu.condonery.logic.commands.property.AddPropertyCommand;
+import seedu.condonery.logic.parser.ArgumentMultimap;
+import seedu.condonery.logic.parser.ArgumentTokenizer;
+import seedu.condonery.logic.parser.Parser;
+import seedu.condonery.logic.parser.ParserUtil;
+import seedu.condonery.logic.parser.Prefix;
 import seedu.condonery.logic.parser.exceptions.ParseException;
 import seedu.condonery.model.property.Address;
 import seedu.condonery.model.property.Name;
@@ -17,17 +22,9 @@ import seedu.condonery.model.property.Property;
 import seedu.condonery.model.tag.Tag;
 
 /**
- * Parses input arguments and creates a new AddCommand object
+ * Parses input arguments and creates a new AddPropertyCommand object
  */
-public class AddCommandParser implements Parser<Command> {
-    private final String directoryArgument;
-
-    /**
-     * Initialises an AddCommandParser with the given directoryArgument.
-     */
-    public AddCommandParser(String directoryArgument) {
-        this.directoryArgument = directoryArgument;
-    }
+public class AddPropertyCommandParser implements Parser<Command> {
 
     /**
      * Parses the given {@code String} of arguments in the context of the AddCommand
@@ -38,23 +35,11 @@ public class AddCommandParser implements Parser<Command> {
         ArgumentMultimap argMultimap =
                 ArgumentTokenizer.tokenize(args, PREFIX_NAME, PREFIX_ADDRESS, PREFIX_TAG);
 
-        if (directoryArgument.equals("-p")) {
-            if (!arePrefixesPresent(argMultimap, PREFIX_NAME, PREFIX_ADDRESS)
-                    || !argMultimap.getPreamble().isEmpty()) {
-                throw new ParseException(String.format(MESSAGE_INVALID_COMMAND_FORMAT, AddPropertyCommand.MESSAGE_USAGE));
-            }
-
-            Name name = ParserUtil.parseName(argMultimap.getValue(PREFIX_NAME).get());
-            Address address = ParserUtil.parseAddress(argMultimap.getValue(PREFIX_ADDRESS).get());
-            Set<Tag> tagList = ParserUtil.parseTags(argMultimap.getAllValues(PREFIX_TAG));
-
-            Property property = new Property(name, address, tagList);
-
-            return new AddPropertyCommand(property);
-        } else { // TODO: Add client
+        if (!arePrefixesPresent(argMultimap, PREFIX_NAME, PREFIX_ADDRESS)
+                || !argMultimap.getPreamble().isEmpty()) {
+            throw new ParseException(String.format(MESSAGE_INVALID_COMMAND_FORMAT, AddPropertyCommand.MESSAGE_USAGE));
         }
 
-        // TODO: This is just a placeholder to allow the function to always have a return statement.
         Name name = ParserUtil.parseName(argMultimap.getValue(PREFIX_NAME).get());
         Address address = ParserUtil.parseAddress(argMultimap.getValue(PREFIX_ADDRESS).get());
         Set<Tag> tagList = ParserUtil.parseTags(argMultimap.getAllValues(PREFIX_TAG));
